backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1

